Update README.md
NUANS (Newly Upgraded Automated Name Search) and Name reservation is same thing. In simple words it is used to reserve a business name prior to the register any corporation or small business. As per the Federal and Provincial regulations to register any named company a NUANS report is required and need to be provided at the time of registration by uploading the report number and date in the online form to proceed with the business registration. Its usually 7 pages long and this report provide the information of conflicts in terms of existing names and trademarks.
Importance of NUANS Report?
This report is a perquisite requirement for the registration of:
•	Corporation
•	Start-ups
•	Not-for-profits. 
Upon generation of the report, the name will be reserved for a period of 90 days.
Why and who need a NUANS Report?

The report is required for:
•	Entrepreneurs, 
•	Lawyers, 
•	Anyone planning to Incorporate at the federal, provincial, or Extra-provincial level